progress

+++ src/main/java/com/mutts_app/repositories/MessageRepository.java
@@ -1,9 +1,7 @@
 package com.mutts_app.repositories;
 
-import com.mutts_app.repositories.pojos.Message;
+import com.mutts_app.pojos.Message;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
@@ -11,10 +9,11 @@ import java.util.List;
 @Repository
 public interface MessageRepository extends JpaRepository<Message, Long> {
 
-    public Message findById(long id);
+    public Message findById(int id);
 
     public Message findFirst1ByChatIdOrderByIdDesc(int chatId);
 
     public List<Message> findMessagesByUserId(int userId);
+
 
 }
